requirements.txt
requests
python-dateutil
dataclasses-json
pydantic
python-dotenv
rich